requirements.txt
aniso8601==9.0.1
astroid==2.4.2
certifi==2020.12.5
chardet==3.0.4
click==8.0.1
clx-sdk-xms==1.0.0
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==2.10
influxdb-client==1.19.0
iso8601==0.1.13
isort==5.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector-python==8.0.25
numpy==1.19.4
opencv-python==4.4.0.46
paho-mqtt==1.5.1
protobuf==3.14.0
PyJWT==1.7.1
pylint==2.6.0
python-cli-ui==0.7.5
python-dateutil==2.8.2
pytz==2020.4
requests==2.25.0
Rx==3.2.0
six==1.15.0
tabulate==0.8.7
toml==0.10.2
twilio==6.49.0
Unidecode==1.1.1
urllib3==1.26.2
Werkzeug==2.0.1
wrapt==1.12.1
